avocado_bot/requirements.txt
aiosqlite==0.20.0
aiohttp==3.9.4
discord.py==2.3.2
python-dotenv==1.0.1
pandas==2.2.2
matplotlib==3.9.0
python-json-logger==2.0.7
pytest==8.2.0
aioresponses==0.7.6